Refactored the course structure to have separate files per each course code.
