+++ src/data/courses.ts
@@ -1,824 +1,16 @@
-import { Course } from "../types";
-
-export const courses: Record<string, Course> = {
-  // ============================================
-  // FIRST YEAR COURSES
-  // ============================================
-
-  // Mathematics - First Year
-  MATH1013: {
-    code: "MATH1013",
-    name: "Mathematics and Applications 1",
-    units: 6,
-    level: 1000,
-    college: "Science",
-    semesters: ["S1", "S2"],
-    prerequisites: [],
-    incompatible: ["MATH1115", "MATH1116"],
-    description: "Calculus, linear algebra, probability basics",
-    type: "foundation",
-  },
-  MATH1014: {
-    code: "MATH1014",
-    name: "Mathematics and Applications 2",
-    units: 6,
-    level: 1000,
-    college: "Science",
-    semesters: ["S1", "S2"],
-    prerequisites: ["MATH1013"],
-    incompatible: ["MATH1116"],
-    description: "Multivariable calculus, ODEs, linear algebra continuation",
-    type: "foundation",
-  },
-  MATH1115: {
-    code: "MATH1115",
-    name: "Advanced Mathematics and Applications 1",
-    units: 6,
-    level: 1000,
-    college: "Science",
-    semesters: ["S1"],
-    prerequisites: [],
-    corequisites: [],
-    assumedKnowledge: "Mathematics Extension 2 or equivalent",
-    incompatible: [],
-    description: "Honours-track calculus and linear algebra",
-    type: "elective",
-  },
-  MATH1116: {
-    code: "MATH1116",
-    name: "Advanced Mathematics and Applications 2",
-    units: 6,
-    level: 1000,
-    college: "Science",
-    semesters: ["S2"],
-    prerequisites: ["MATH1115"],
-    prerequisiteAlternatives: [["MATH1115"], ["MATH1013"]],
-    incompatible: ["MATH1014"],
-    description: "Honours-track multivariable calculus and ODEs",
-    type: "foundation",
-  },
-
-  // Physics - First Year
-  PHYS1101: {
-    code: "PHYS1101",
-    name: "Physics 1",
-    units: 6,
-    level: 1000,
-    college: "Science",
-    semesters: ["S1", "S2"],
-    prerequisites: [],
-    assumedKnowledge: "HSC Physics or equivalent",
-    incompatible: ["PHYS1001"],
-    description:
-      "Mechanics, circuits, electromagnetism introduction. Recommended for engineering students.",
-    type: "foundation",
-  },
-  PHYS1201: {
-    code: "PHYS1201",
-    name: "Physics 2",
-    units: 6,
-    level: 1000,
-    college: "Science",
-    semesters: ["S2"],
-    prerequisites: ["PHYS1101"],
-    prerequisiteAlternatives: [["PHYS1101"], ["PHYS1001"]],
-    description: "Waves, optics, thermodynamics, Fourier analysis introduction",
-    type: "elective",
-  },
-
-  // Computing - First Year
-  COMP1100: {
-    code: "COMP1100",
-    name: "Programming as Problem Solving",
-    units: 6,
-    level: 1000,
-    college: "Engineering & Computing",
-    semesters: ["S1", "S2"],
-    prerequisites: [],
-    incompatible: ["COMP1130"],
-    description: "Functional programming in Haskell, computational thinking",
-    type: "foundation",
-  },
-  COMP1110: {
-    code: "COMP1110",
-    name: "Structured Programming",
-    units: 6,
-    level: 1000,
-    college: "Engineering & Computing",
-    semesters: ["S1", "S2"],
-    prerequisites: ["COMP1100"],
-    prerequisiteAlternatives: [["COMP1100"], ["COMP1130"]],
-    incompatible: ["COMP1140"],
-    description: "Object-oriented programming in Java",
-    type: "elective",
-  },
-  COMP1130: {
-    code: "COMP1130",
-    name: "Programming as Problem Solving (Advanced)",
-    units: 6,
-    level: 1000,
-    college: "Engineering & Computing",
-    semesters: ["S1"],
-    prerequisites: [],
-    incompatible: ["COMP1100"],
-    description: "Advanced functional programming",
-    type: "elective",
-  },
-
-  // Engineering - First Year
-  ENGN1211: {
-    code: "ENGN1211",
-    name: "Discovering Engineering",
-    units: 6,
-    level: 1000,
-    college: "Engineering",
-    semesters: ["S1"],
-    prerequisites: [],
-    description: "Introduction to engineering disciplines, team projects",
-    type: "foundation",
-  },
-  ENGN1217: {
-    code: "ENGN1217",
-    name: "Introduction to Mechanics",
-    units: 6,
-    level: 1000,
-    college: "Engineering",
-    semesters: ["S1", "S2"],
-    prerequisites: [],
-    corequisites: ["MATH1013"],
-    description: "Statics, dynamics, materials introduction",
-    type: "foundation",
-  },
-  ENGN1218: {
-    code: "ENGN1218",
-    name: "Introduction to Electronics",
-    units: 6,
-    level: 1000,
-    college: "Engineering",
-    semesters: ["S1", "S2"],
-    prerequisites: [],
-    corequisites: ["MATH1013"],
-    description: "Circuit analysis, basic electronics, lab skills",
-    type: "foundation",
-    majorRelevance: ["ECSY-MAJ"],
-  },
-
-  // ============================================
-  // SECOND YEAR COURSES
-  // ============================================
-
-  // Mathematics - Second Year
-  MATH2305: {
-    code: "MATH2305",
-    name: "Applied Mathematics I",
-    units: 6,
-    level: 2000,
-    college: "Science",
-    semesters: ["S1"],
-    prerequisites: ["MATH1014"],
-    prerequisiteAlternatives: [["MATH1014"], ["MATH1116"]],
-    description:
-      "ODEs, vector calculus (Green's, Stokes', divergence theorems), numerical methods in MATLAB",
-    type: "elective",
-    recommended: "Essential for engineering applications",
-  },
-  MATH2306: {
-    code: "MATH2306",
-    name: "Applied Mathematics II",
-    units: 6,
-    level: 2000,
-    college: "Science",
-    semesters: ["S2"],
-    prerequisites: ["MATH2305"],
-    description: "PDEs, Fourier series and transforms, boundary value problems",
-    type: "elective",
-    recommended: "Critical for electromagnetic wave problems",
-  },
-  MATH2320: {
-    code: "MATH2320",
-    name: "Analysis 1: Metric Spaces and Applications",
-    units: 6,
-    level: 2000,
-    college: "Science",
-    semesters: ["S1"],
-    prerequisites: ["MATH1116"],
-    prerequisiteAlternatives: [["MATH1116"], ["MATH1014"]],
-    description: "Honours-track real analysis, metric spaces",
-    type: "elective",
-  },
-  MATH2322: {
-    code: "MATH2322",
-    name: "Algebra 1",
-    units: 6,
-    level: 2000,
-    college: "Science",
-    semesters: ["S2"],
-    prerequisites: ["MATH1014"],
-    prerequisiteAlternatives: [["MATH1014"], ["MATH1116"]],
-    description: "Abstract algebra, groups, rings",
-    type: "elective",
-  },
-
-  // Physics - Second Year
-  PHYS2013: {
-    code: "PHYS2013",
-    name: "Quantum Mechanics",
-    units: 6,
-    level: 2000,
-    college: "Science",
-    semesters: ["S1"],
-    prerequisites: ["PHYS1201", "MATH1014"],
-    description: "Introduction to quantum mechanics, wavefunctions, operators",
-    type: "elective",
-  },
-  PHYS2016: {
-    code: "PHYS2016",
-    name: "Electromagnetism",
-    units: 6,
-    level: 2000,
-    college: "Science",
-    semesters: ["S1"],
-    prerequisites: ["PHYS1101", "MATH1014"],
-    description:
-      "Maxwell's equations, electromagnetic waves, vector calculus applications",
-    type: "elective",
-    recommended: "Critical first-principles course for RF electronics",
-  },
-  PHYS2020: {
-    code: "PHYS2020",
-    name: "Thermal and Statistical Physics",
-    units: 6,
-    level: 2000,
-    college: "Science",
-    semesters: ["S2"],
-    prerequisites: ["PHYS1201", "MATH1014"],
-    description: "Thermodynamics, statistical mechanics",
-    type: "elective",
-  },
-  PHYS2201: {
-    code: "PHYS2201",
-    name: "Classical Mechanics",
-    units: 6,
-    level: 2000,
-    college: "Science",
-    semesters: ["S2"],
-    prerequisites: ["PHYS1101", "MATH1014"],
-    description: "Lagrangian and Hamiltonian mechanics",
-    type: "elective",
-  },
-
-  // Computing - Second Year
-  COMP2100: {
-    code: "COMP2100",
-    name: "Software Design Methodologies",
-    units: 6,
-    level: 2000,
-    college: "Engineering & Computing",
-    semesters: ["S1", "S2"],
-    prerequisites: ["COMP1110"],
-    prerequisiteAlternatives: [["COMP1110"], ["COMP1140"]],
-    description: "Software engineering, design patterns, testing",
-    type: "elective",
-  },
-  COMP2310: {
-    code: "COMP2310",
-    name: "Systems, Networks and Concurrency",
-    units: 6,
-    level: 2000,
-    college: "Engineering & Computing",
-    semesters: ["S2"],
-    prerequisites: ["COMP1110"],
-    description: "Concurrent programming, networking basics",
-    type: "elective",
-  },
-
-  // Engineering - Second Year Core
-  ENGN2217: {
-    code: "ENGN2217",
-    name: "Mechanical Systems and Design",
-    units: 6,
-    level: 2000,
-    college: "Engineering",
-    semesters: ["S1"],
-    prerequisites: ["ENGN1217", "MATH1013"],
-    description: "Solid mechanics, machine elements, mechanical design",
-    type: "core",
-  },
-  ENGN2218: {
-    code: "ENGN2218",
-    name: "Electronic Systems and Design",
-    units: 6,
-    level: 2000,
-    college: "Engineering",
-    semesters: ["S1"],
-    prerequisites: ["ENGN1218", "MATH1013"],
-    description: "BJT/FET amplifiers, op-amps, digital logic, PCB design",
-    type: "core",
-    majorRelevance: ["ECSY-MAJ"],
-  },
-  ENGN2219: {
-    code: "ENGN2219",
-    name: "Computer Systems and Organisation",
-    units: 6,
-    level: 2000,
-    college: "Engineering",
-    semesters: ["S1"],
-    prerequisites: ["COMP1100"],
-    description: "Computer architecture, assembly, memory systems",
-    type: "core",
-    note: "Prerequisite for HPC courses",
-  },
-  ENGN2222: {
-    code: "ENGN2222",
-    name: "Engineering Thermodynamics",
-    units: 6,
-    level: 2000,
-    college: "Engineering",
-    semesters: ["S2"],
-    prerequisites: ["MATH1014", "PHYS1101"],
-    description: "Thermodynamic cycles, heat transfer",
-    type: "core",
-  },
-  ENGN2228: {
-    code: "ENGN2228",
-    name: "Signals and Systems",
-    units: 6,
-    level: 2000,
-    college: "Engineering",
-    semesters: ["S2"],
-    prerequisites: ["MATH1013"],
-    description:
-      "Fourier analysis, LTI systems, Laplace transforms, frequency response",
-    type: "core",
-    majorRelevance: ["ECSY-MAJ"],
-  },
-
-  // Engineering - Second Year Professional Core
-  ENGN2300: {
-    code: "ENGN2300",
-    name: "Engineering Design 2",
-    units: 6,
-    level: 2000,
-    college: "Engineering",
-    semesters: ["S1"],
-    prerequisites: ["ENGN1211"],
-    description: "Team-based design project, project management",
-    type: "professionalCore",
-  },
-  ENGN2301: {
-    code: "ENGN2301",
-    name: "Engineering Design 3",
-    units: 6,
-    level: 2000,
-    college: "Engineering",
-    semesters: ["S2"],
-    prerequisites: ["ENGN2300"],
-    description: "Systems engineering, requirements analysis",
-    type: "professionalCore",
-  },
-
-  // ============================================
-  // THIRD YEAR COURSES
-  // ============================================
-
-  // Mathematics - Third Year
-
-  MATH3320: {
-    code: "MATH3320",
-    name: "Advanced Analysis 2",
-    units: 6,
-    level: 3000,
-    college: "CSS",
-    semesters: ["S1"],
-    prerequisites: ["MATH2320"],
-    description:
-      "Explain the fundamental concepts of advanced analysis such as Lebesgue measure and integration and Hilbert space theory and their role in modern mathematics and applied contexts",
-    type: "elective",
-    recommended: "Improved Analysis techniques",
-  },
-
-  MATH3511: {
-    code: "MATH3511",
-    name: "Scientific Computing",
-    units: 6,
-    level: 3000,
-    college: "Science",
-    semesters: ["S1"],
-    prerequisites: ["MATH2305"],
-    prerequisiteAlternatives: [["MATH2305"], ["MATH2405"]],
-    description:
-      "Numerical methods for ODEs and PDEs, iterative methods, stability analysis",
-    type: "elective",
-    recommended: "Core numerical methods course",
-  },
-  MATH3512: {
-    code: "MATH3512",
-    name: "Matrix Computations",
-    units: 6,
-    level: 3000,
-    college: "Science",
-    semesters: ["S2"],
-    semesterPattern: "odd_years_only",
-    offeredYears: [2025, 2027, 2029],
-    prerequisites: ["MATH2305"],
-    prerequisiteAlternatives: [["MATH2305"], ["MATH1014"]],
-    description:
-      "Numerical linear algebra, matrix factorizations, eigenvalue algorithms, SVD",
-    type: "elective",
-    recommended:
-      "Essential for large-scale simulations. OFFERED ODD YEARS ONLY.",
-  },
-  MATH3514: {
-    code: "MATH3514",
-    name: "Numerical Optimisation",
-    units: 6,
-    level: 3000,
-    college: "Science",
-    semesters: ["S2"],
-    semesterPattern: "even_years_only",
-    offeredYears: [2026, 2028, 2030],
-    prerequisites: ["MATH2305"],
-    prerequisiteAlternatives: [["MATH2305"], ["MATH1014"]],
-    description:
-      "Gradient descent, Newton methods, constrained optimization, convex optimization",
-    type: "elective",
-    recommended:
-      "Critical for antenna design and control systems. OFFERED EVEN YEARS ONLY.",
-  },
-  MATH3029: {
-    code: "MATH3029",
-    name: "Probability Modelling and Applications",
-    units: 6,
-    level: 3000,
-    college: "Science",
-    semesters: ["S1"],
-    prerequisites: ["MATH1014"],
-    description: "Stochastic processes, Markov chains, queueing theory",
-    type: "elective",
-  },
-  MATH3325: {
-    code: "MATH3325",
-    name: "Functional Analysis",
-    units: 6,
-    level: 3000,
-    college: "Science",
-    semesters: ["S2"],
-    prerequisites: ["MATH2320"],
-    description: "Banach and Hilbert spaces, bounded operators",
-    type: "elective",
-  },
-  MATH3228: {
-    code: "MATH3228",
-    name: "Complex Analysis",
-    units: 6,
-    level: 3000,
-    college: "Science",
-    semesters: ["S1"],
-    prerequisites: ["MATH2320"],
-    prerequisiteAlternatives: [["MATH2320"], ["MATH2305"]],
-    description: "Complex functions, contour integration, residue theorem",
-    type: "elective",
-  },
-
-  // Physics - Third Year
-  PHYS3032: {
-    code: "PHYS3032",
-    name: "Solid State Physics",
-    units: 6,
-    level: 3000,
-    college: "Science",
-    semesters: ["S1"],
-    prerequisites: ["PHYS2013"],
-    description:
-      "Crystal structure, band theory, semiconductors, device physics",
-    type: "elective",
-    recommended: "Essential for understanding RF component behavior",
-  },
-  PHYS3101: {
-    code: "PHYS3101",
-    name: "Advanced Quantum Mechanics",
-    units: 6,
-    level: 3000,
-    college: "Science",
-    semesters: ["S1"],
-    prerequisites: ["PHYS2013", "MATH2306"],
-    description: "Perturbation theory, scattering, identical particles",
-    type: "elective",
-  },
-  PHYS3102: {
-    code: "PHYS3102",
-    name: "Advanced Electromagnetism",
-    units: 6,
-    level: 3000,
-    college: "Science",
-    semesters: ["S2"],
-    prerequisites: ["PHYS2016", "MATH2306"],
-    description:
-      "Radiation from currents, waveguides, transmission lines, covariant formulation",
-    type: "elective",
-    recommended: "Advanced EM theory foundation for RF technologies",
-  },
-
-  // Computing - Third Year
-  COMP3320: {
-    code: "COMP3320",
-    name: "High Performance Scientific Computation",
-    units: 6,
-    level: 3000,
-    college: "Engineering & Computing",
-    semesters: ["S1"],
-    prerequisites: ["COMP1110", "ENGN2219"],
-    prerequisiteAlternatives: [["COMP1110", "ENGN2219"], ["COMP2310"]],
-    description:
-      "Parallel programming, OpenMP, performance optimization, C/C++ for HPC",
-    type: "elective",
-    recommended: "Core HPC course",
-  },
-  COMP3600: {
-    code: "COMP3600",
-    name: "Algorithms",
-    units: 6,
-    level: 3000,
-    college: "Engineering & Computing",
-    semesters: ["S1"],
-    prerequisites: ["COMP1110", "MATH1014"],
-    description: "Algorithm design and analysis, complexity theory",
-    type: "elective",
-  },
-
-  // Engineering - Third Year Major
-  ENGN3226: {
-    code: "ENGN3226",
-    name: "Digital Communications",
-    units: 6,
-    level: 3000,
-    college: "Engineering",
-    semesters: ["S1"],
-    prerequisites: ["ENGN2228"],
-    description: "Modulation, channel coding, digital transmission systems",
-    type: "major",
-    majorRelevance: ["ECSY-MAJ"],
-    note: "Being replaced by ENGN3350 from 2027",
-  },
-  ENGN3331: {
-    code: "ENGN3331",
-    name: "System Dynamics",
-    units: 6,
-    level: 3000,
-    college: "Engineering",
-    semesters: ["S1"],
-    prerequisites: ["MATH1013"],
-    description:
-      "System dynamics is the study of the response of engineering systems with changing time",
-    type: "engnElective",
-  },
-  ENGN3338: {
-    code: "ENGN3338",
-    name: "Aerodynamics",
-    units: 6,
-    level: 3000,
-    college: "Engineering",
-    semesters: ["S1"],
-    prerequisites: ["ENGN2222"],
-    description: "low speed and high-speed aerodynamics of wings and bodies",
-    type: "engnElective",
-  },
-  ENGN4339: {
-    code: "ENGN4339",
-    name: "Aerospace Instrumentation and Avionics",
-    units: 6,
-    level: 3000,
-    college: "Engineering",
-    semesters: ["S2"],
-    prerequisites: ["ENGN3338"],
-    description:
-      "developing microcontroller-based instrumentation for aerospace applications",
-    type: "engnElective",
-  },
-  ENGN3350: {
-    code: "ENGN3350",
-    name: "Communications Systems",
-    units: 6,
-    level: 3000,
-    college: "Engineering",
-    semesters: ["S1"],
-    prerequisites: ["ENGN2228"],
-    description: "Modern communications systems, replacement for ENGN3226",
-    type: "major",
-    majorRelevance: ["ECSY-MAJ"],
-    note: "Replaces ENGN3226 from 2027",
-  },
-  ENGN3539: {
-    code: "ENGN3539",
-    name: "Computer Networks",
-    units: 6,
-    level: 3000,
-    college: "Engineering",
-    semesters: ["S1"],
-    prerequisites: ["ENGN2219"],
-    description:
-      "Network protocols, TCP/IP, wireless networks, network security",
-    type: "major",
-    majorRelevance: ["ECSY-MAJ"],
-  },
-
-  // Engineering - Third Year Professional Core
-  ENGN3300: {
-    code: "ENGN3300",
-    name: "Engineering Design 4A",
-    units: 6,
-    level: 3000,
-    college: "Engineering",
-    semesters: ["S1"],
-    prerequisites: ["ENGN2301"],
-    description: "Advanced systems engineering, stakeholder management",
-    type: "professionalCore",
-  },
-  ENGN3301: {
-    code: "ENGN3301",
-    name: "Engineering Design 4B",
-    units: 6,
-    level: 3000,
-    college: "Engineering",
-    semesters: ["S2"],
-    prerequisites: ["ENGN3300"],
-    description: "Systems integration, verification and validation",
-    type: "professionalCore",
-  },
-
-  // Industry Experience
-  ENGN3100: {
-    code: "ENGN3100",
-    name: "Professional Engineering",
-    units: 0,
-    level: 3000,
-    college: "Engineering",
-    semesters: ["S1", "S2", "Summer"],
-    prerequisites: [],
-    description: "60 days industry work experience requirement",
-    type: "industryExperience",
-  },
-
-  // ============================================
-  // FOURTH YEAR COURSES
-  // ============================================
-
-  // Computing - Fourth Year
-  COMP4300: {
-    code: "COMP4300",
-    name: "Parallel Systems",
-    units: 6,
-    level: 4000,
-    college: "Engineering & Computing",
-    semesters: ["S1"],
-    prerequisites: ["COMP3320"],
-    prerequisiteAlternatives: [["COMP3320"], ["COMP2310", "COMP2420"]],
-    description: "MPI, CUDA, GPU computing, access to NCI Gadi supercomputer",
-    type: "elective",
-    recommended: "Advanced HPC for large-scale EM simulations",
-  },
-
-  // Engineering - Fourth Year Major
-  ENGN4213: {
-    code: "ENGN4213",
-    name: "Digital Systems and Microprocessors",
-    units: 6,
-    level: 4000,
-    college: "Engineering",
-    semesters: ["S1"],
-    prerequisites: ["ENGN2218", "ENGN2219"],
-    description: "FPGA design, embedded systems, HDL programming",
-    type: "major",
-    majorRelevance: ["ECSY-MAJ"],
-  },
-  ENGN4536: {
-    code: "ENGN4536",
-    name: "Wireless Communications",
-    units: 6,
-    level: 4000,
-    college: "Engineering",
-    semesters: ["S2"],
-    prerequisites: ["ENGN3226"],
-    prerequisiteAlternatives: [["ENGN3226"], ["ENGN3350"]],
-    description:
-      "RF channel modeling, MIMO systems, antenna fundamentals, 5G technologies",
-    type: "major",
-    majorRelevance: ["ECSY-MAJ"],
-  },
-  ENGN4537: {
-    code: "ENGN4537",
-    name: "Digital Signal Processing",
-    units: 6,
-    level: 4000,
-    college: "Engineering",
-    semesters: ["S2"],
-    prerequisites: ["ENGN2228"],
-    description: "Z-transforms, digital filter design, FFT, DSP implementation",
-    type: "major",
-    majorRelevance: ["ECSY-MAJ"],
-    note: "Being replaced by ENGN4430 from 2027",
-  },
-  ENGN4430: {
-    code: "ENGN4430",
-    name: "Signal Processing and Machine Learning",
-    units: 6,
-    level: 4000,
-    college: "Engineering",
-    semesters: ["S2"],
-    prerequisites: ["ENGN2228"],
-    description:
-      "Modern signal processing with ML integration, replacement for ENGN4537",
-    type: "major",
-    majorRelevance: ["ECSY-MAJ"],
-    note: "Replaces ENGN4537 from 2027",
-  },
-  ENGN4625: {
-    code: "ENGN4625",
-    name: "Power Systems and Power Electronics",
-    units: 6,
-    level: 4000,
-    college: "Engineering",
-    semesters: ["S2"],
-    prerequisites: ["ENGN2218", "ENGN2228"],
-    description: "Power conversion, inverters, power system analysis",
-    type: "major",
-    majorRelevance: ["ECSY-MAJ"],
-  },
-
-  // Capstone
-  ENGN4300: {
-    code: "ENGN4300",
-    name: "Engineering Capstone Project",
-    units: 12,
-    level: 4000,
-    college: "Engineering",
-    semesters: ["S1", "S2"],
-    semesterSpan: 2,
-    prerequisites: ["ENGN3301"],
-    description: "Year-long team project with industry partner",
-    type: "capstone",
-    honoursWeight: 0.4,
-  },
-
-  // Additional ENGN Electives
-  ENGN3221: {
-    code: "ENGN3221",
-    name: "Analogue Electronics",
-    units: 6,
-    level: 3000,
-    college: "Engineering",
-    semesters: ["S1"],
-    prerequisites: ["ENGN2218"],
-    description: "Advanced analog circuit design, RF amplifiers",
-    type: "engnElective",
-    recommended: "Valuable for RF hardware design",
-  },
-  ENGN3230: {
-    code: "ENGN3230",
-    name: "Control Systems",
-    units: 6,
-    level: 3000,
-    college: "Engineering",
-    semesters: ["S2"],
-    prerequisites: ["ENGN2228"],
-    description: "Classical and modern control theory, feedback systems",
-    type: "engnElective",
-  },
-  ENGN4524: {
-    code: "ENGN4524",
-    name: "Computer Vision and Image Processing",
-    units: 6,
-    level: 4000,
-    college: "Engineering",
-    semesters: ["S1"],
-    prerequisites: ["ENGN2228"],
-    description:
-      "Image processing, feature extraction, computer vision algorithms",
-    type: "engnElective",
-  },
-  ENGN4528: {
-    code: "ENGN4528",
-    name: "Computer Vision",
-    units: 6,
-    level: 4000,
-    college: "Engineering",
-    semesters: ["S2"],
-    prerequisites: ["ENGN4524"],
-    description: "Advanced computer vision, deep learning for vision",
-    type: "engnElective",
-  },
-};
-
-export const courseList = Object.values(courses);
-
-export const getCoursesByType = (type: Course["type"]): Course[] => {
-  return courseList.filter((c) => c.type === type);
-};
-
-export const getCoursesByLevel = (level: number): Course[] => {
-  return courseList.filter((c) => c.level === level);
-};
-
-export const getCourseBySemester = (semester: "S1" | "S2"): Course[] => {
-  return courseList.filter((c) => c.semesters.includes(semester));
-};
+// Re-export everything from the new modular structure
+// This maintains backwards compatibility with existing imports
+export {
+  courses,
+  courseList,
+  getCoursesByType,
+  getCoursesByLevel,
+  getCourseBySemester,
+  getCoursesByPrefix,
+  getCoursePrefixes,
+  courseModules,
+  COMP_COURSES,
+  ENGN_COURSES,
+  MATH_COURSES,
+  PHYS_COURSES,
+} from "./courses/index";
