+++ src/data/courses/ENGN.ts
@@ -1,0 +1,341 @@
+import { Course } from "../../types";
+
+export const ENGN_COURSES: Record<string, Course> = {
+  ENGN1211: {
+    code: "ENGN1211",
+    name: "Discovering Engineering",
+    units: 6,
+    level: 1000,
+    college: "Engineering",
+    semesters: ["S1"],
+    prerequisites: [],
+    description: "Introduction to engineering disciplines, team projects",
+    type: "foundation",
+  },
+  ENGN1217: {
+    code: "ENGN1217",
+    name: "Introduction to Mechanics",
+    units: 6,
+    level: 1000,
+    college: "Engineering",
+    semesters: ["S1", "S2"],
+    prerequisites: [],
+    corequisites: ["MATH1013"],
+    description: "Statics, dynamics, materials introduction",
+    type: "foundation",
+  },
+  ENGN1218: {
+    code: "ENGN1218",
+    name: "Introduction to Electronics",
+    units: 6,
+    level: 1000,
+    college: "Engineering",
+    semesters: ["S1", "S2"],
+    prerequisites: [],
+    corequisites: ["MATH1013"],
+    description: "Circuit analysis, basic electronics, lab skills",
+    type: "foundation",
+    majorRelevance: ["ECSY-MAJ"],
+  },
+  ENGN2217: {
+    code: "ENGN2217",
+    name: "Mechanical Systems and Design",
+    units: 6,
+    level: 2000,
+    college: "Engineering",
+    semesters: ["S1"],
+    prerequisites: ["ENGN1217", "MATH1013"],
+    description: "Solid mechanics, machine elements, mechanical design",
+    type: "core",
+  },
+  ENGN2218: {
+    code: "ENGN2218",
+    name: "Electronic Systems and Design",
+    units: 6,
+    level: 2000,
+    college: "Engineering",
+    semesters: ["S1"],
+    prerequisites: ["ENGN1218", "MATH1013"],
+    description: "BJT/FET amplifiers, op-amps, digital logic, PCB design",
+    type: "core",
+    majorRelevance: ["ECSY-MAJ"],
+  },
+  ENGN2219: {
+    code: "ENGN2219",
+    name: "Computer Systems and Organisation",
+    units: 6,
+    level: 2000,
+    college: "Engineering",
+    semesters: ["S1"],
+    prerequisites: ["COMP1100"],
+    description: "Computer architecture, assembly, memory systems",
+    type: "core",
+    note: "Prerequisite for HPC courses",
+  },
+  ENGN2222: {
+    code: "ENGN2222",
+    name: "Engineering Thermodynamics",
+    units: 6,
+    level: 2000,
+    college: "Engineering",
+    semesters: ["S2"],
+    prerequisites: ["MATH1014", "PHYS1101"],
+    description: "Thermodynamic cycles, heat transfer",
+    type: "core",
+  },
+  ENGN2228: {
+    code: "ENGN2228",
+    name: "Signals and Systems",
+    units: 6,
+    level: 2000,
+    college: "Engineering",
+    semesters: ["S2"],
+    prerequisites: ["MATH1013"],
+    description:
+      "Fourier analysis, LTI systems, Laplace transforms, frequency response",
+    type: "core",
+    majorRelevance: ["ECSY-MAJ"],
+  },
+  ENGN2300: {
+    code: "ENGN2300",
+    name: "Engineering Design 2",
+    units: 6,
+    level: 2000,
+    college: "Engineering",
+    semesters: ["S1"],
+    prerequisites: ["ENGN1211"],
+    description: "Team-based design project, project management",
+    type: "professionalCore",
+  },
+  ENGN2301: {
+    code: "ENGN2301",
+    name: "Engineering Design 3",
+    units: 6,
+    level: 2000,
+    college: "Engineering",
+    semesters: ["S2"],
+    prerequisites: ["ENGN2300"],
+    description: "Systems engineering, requirements analysis",
+    type: "professionalCore",
+  },
+  ENGN3100: {
+    code: "ENGN3100",
+    name: "Professional Engineering",
+    units: 0,
+    level: 3000,
+    college: "Engineering",
+    semesters: ["S1", "S2", "Summer"],
+    prerequisites: [],
+    description: "60 days industry work experience requirement",
+    type: "industryExperience",
+  },
+  
+  ENGN3226: {
+    code: "ENGN3226",
+    name: "Digital Communications",
+    units: 6,
+    level: 3000,
+    college: "Engineering",
+    semesters: ["S1"],
+    prerequisites: ["ENGN2228"],
+    description: "Modulation, channel coding, digital transmission systems",
+    type: "major",
+    majorRelevance: ["ECSY-MAJ"],
+    note: "Being replaced by ENGN3350 from 2027",
+  },
+  ENGN3230: {
+    code: "ENGN3230",
+    name: "Control Systems",
+    units: 6,
+    level: 3000,
+    college: "Engineering",
+    semesters: ["S2"],
+    prerequisites: ["ENGN2228"],
+    description: "Classical and modern control theory, feedback systems",
+    type: "engnElective",
+  },
+  ENGN3300: {
+    code: "ENGN3300",
+    name: "Engineering Design 4A",
+    units: 6,
+    level: 3000,
+    college: "Engineering",
+    semesters: ["S1"],
+    prerequisites: ["ENGN2301"],
+    description: "Advanced systems engineering, stakeholder management",
+    type: "professionalCore",
+  },
+  ENGN3301: {
+    code: "ENGN3301",
+    name: "Engineering Design 4B",
+    units: 6,
+    level: 3000,
+    college: "Engineering",
+    semesters: ["S2"],
+    prerequisites: ["ENGN3300"],
+    description: "Systems integration, verification and validation",
+    type: "professionalCore",
+  },
+  ENGN3331: {
+    code: "ENGN3331",
+    name: "System Dynamics",
+    units: 6,
+    level: 3000,
+    college: "Engineering",
+    semesters: ["S1"],
+    prerequisites: ["MATH1013"],
+    description:
+      "System dynamics is the study of the response of engineering systems with changing time",
+    type: "engnElective",
+  },
+  ENGN3338: {
+    code: "ENGN3338",
+    name: "Aerodynamics",
+    units: 6,
+    level: 3000,
+    college: "Engineering",
+    semesters: ["S1"],
+    prerequisites: ["ENGN2222"],
+    description: "low speed and high-speed aerodynamics of wings and bodies",
+    type: "engnElective",
+  },
+  ENGN3350: {
+    code: "ENGN3350",
+    name: "Communications Systems",
+    units: 6,
+    level: 3000,
+    college: "Engineering",
+    semesters: ["S1"],
+    prerequisites: ["ENGN2228"],
+    description: "Modern communications systems, replacement for ENGN3226",
+    type: "major",
+    majorRelevance: ["ECSY-MAJ"],
+    note: "Replaces ENGN3226 from 2027",
+  },
+  ENGN3539: {
+    code: "ENGN3539",
+    name: "Computer Networks",
+    units: 6,
+    level: 3000,
+    college: "Engineering",
+    semesters: ["S1"],
+    prerequisites: ["ENGN2219"],
+    description:
+      "Network protocols, TCP/IP, wireless networks, network security",
+    type: "major",
+    majorRelevance: ["ECSY-MAJ"],
+  },
+  ENGN4213: {
+    code: "ENGN4213",
+    name: "Digital Systems and Microprocessors",
+    units: 6,
+    level: 4000,
+    college: "Engineering",
+    semesters: ["S1"],
+    prerequisites: ["ENGN2218", "ENGN2219"],
+    description: "FPGA design, embedded systems, HDL programming",
+    type: "major",
+    majorRelevance: ["ECSY-MAJ"],
+  },
+  ENGN4300: {
+    code: "ENGN4300",
+    name: "Engineering Capstone Project",
+    units: 12,
+    level: 4000,
+    college: "Engineering",
+    semesters: ["S1", "S2"],
+    semesterSpan: 2,
+    prerequisites: ["ENGN3301"],
+    description: "Year-long team project with industry partner",
+    type: "capstone",
+    honoursWeight: 0.4,
+  },
+  ENGN4339: {
+    code: "ENGN4339",
+    name: "Aerospace Instrumentation and Avionics",
+    units: 6,
+    level: 4000,
+    college: "Engineering",
+    semesters: ["S2"],
+    prerequisites: ["ENGN3338"],
+    description:
+      "developing microcontroller-based instrumentation for aerospace applications",
+    type: "engnElective",
+  },
+  ENGN4430: {
+    code: "ENGN4430",
+    name: "Signal Processing and Machine Learning",
+    units: 6,
+    level: 4000,
+    college: "Engineering",
+    semesters: ["S2"],
+    prerequisites: ["ENGN2228"],
+    description:
+      "Modern signal processing with ML integration, replacement for ENGN4537",
+    type: "major",
+    majorRelevance: ["ECSY-MAJ"],
+    note: "Replaces ENGN4537 from 2027",
+  },
+  ENGN4524: {
+    code: "ENGN4524",
+    name: "Computer Vision and Image Processing",
+    units: 6,
+    level: 4000,
+    college: "Engineering",
+    semesters: ["S1"],
+    prerequisites: ["ENGN2228"],
+    description:
+      "Image processing, feature extraction, computer vision algorithms",
+    type: "engnElective",
+  },
+  ENGN4528: {
+    code: "ENGN4528",
+    name: "Computer Vision",
+    units: 6,
+    level: 4000,
+    college: "Engineering",
+    semesters: ["S2"],
+    prerequisites: ["ENGN4524"],
+    description: "Advanced computer vision, deep learning for vision",
+    type: "engnElective",
+  },
+  ENGN4536: {
+    code: "ENGN4536",
+    name: "Wireless Communications",
+    units: 6,
+    level: 4000,
+    college: "Engineering",
+    semesters: ["S2"],
+    prerequisites: ["ENGN3226"],
+    prerequisiteAlternatives: [["ENGN3226"], ["ENGN3350"]],
+    description:
+      "RF channel modeling, MIMO systems, antenna fundamentals, 5G technologies",
+    type: "major",
+    majorRelevance: ["ECSY-MAJ"],
+  },
+  ENGN4537: {
+    code: "ENGN4537",
+    name: "Digital Signal Processing",
+    units: 6,
+    level: 4000,
+    college: "Engineering",
+    semesters: ["S2"],
+    prerequisites: ["ENGN2228"],
+    description: "Z-transforms, digital filter design, FFT, DSP implementation",
+    type: "major",
+    majorRelevance: ["ECSY-MAJ"],
+    note: "Being replaced by ENGN4430 from 2027",
+  },
+  ENGN4625: {
+    code: "ENGN4625",
+    name: "Power Systems and Power Electronics",
+    units: 6,
+    level: 4000,
+    college: "Engineering",
+    semesters: ["S2"],
+    prerequisites: ["ENGN2218", "ENGN2228"],
+    description: "Power conversion, inverters, power system analysis",
+    type: "major",
+    majorRelevance: ["ECSY-MAJ"],
+  },
+};
