+++ src/data/courses/MATH.ts
@@ -1,0 +1,200 @@
+import { Course } from "../../types";
+
+export const MATH_COURSES: Record<string, Course> = {
+  MATH1013: {
+    code: "MATH1013",
+    name: "Mathematics and Applications 1",
+    units: 6,
+    level: 1000,
+    college: "Science",
+    semesters: ["S1", "S2"],
+    prerequisites: [],
+    incompatible: ["MATH1115", "MATH1116"],
+    description: "Calculus, linear algebra, probability basics",
+    type: "foundation",
+  },
+  MATH1014: {
+    code: "MATH1014",
+    name: "Mathematics and Applications 2",
+    units: 6,
+    level: 1000,
+    college: "Science",
+    semesters: ["S1", "S2"],
+    prerequisites: ["MATH1013"],
+    incompatible: ["MATH1116"],
+    description: "Multivariable calculus, ODEs, linear algebra continuation",
+    type: "foundation",
+  },
+  MATH1115: {
+    code: "MATH1115",
+    name: "Advanced Mathematics and Applications 1",
+    units: 6,
+    level: 1000,
+    college: "Science",
+    semesters: ["S1"],
+    prerequisites: [],
+    corequisites: [],
+    assumedKnowledge: "Mathematics Extension 2 or equivalent",
+    incompatible: [],
+    description: "Honours-track calculus and linear algebra",
+    type: "elective",
+  },
+  MATH1116: {
+    code: "MATH1116",
+    name: "Advanced Mathematics and Applications 2",
+    units: 6,
+    level: 1000,
+    college: "Science",
+    semesters: ["S2"],
+    prerequisites: ["MATH1115"],
+    prerequisiteAlternatives: [["MATH1115"], ["MATH1013"]],
+    incompatible: ["MATH1014"],
+    description: "Honours-track multivariable calculus and ODEs",
+    type: "foundation",
+  },
+  MATH2305: {
+    code: "MATH2305",
+    name: "Applied Mathematics I",
+    units: 6,
+    level: 2000,
+    college: "Science",
+    semesters: ["S1"],
+    prerequisites: ["MATH1014"],
+    prerequisiteAlternatives: [["MATH1014"], ["MATH1116"]],
+    description:
+      "ODEs, vector calculus (Green's, Stokes', divergence theorems), numerical methods in MATLAB",
+    type: "elective",
+    recommended: "Essential for engineering applications",
+  },
+  MATH2306: {
+    code: "MATH2306",
+    name: "Applied Mathematics II",
+    units: 6,
+    level: 2000,
+    college: "Science",
+    semesters: ["S2"],
+    prerequisites: ["MATH2305"],
+    description: "PDEs, Fourier series and transforms, boundary value problems",
+    type: "elective",
+    recommended: "Critical for electromagnetic wave problems",
+  },
+  MATH2320: {
+    code: "MATH2320",
+    name: "Analysis 1: Metric Spaces and Applications",
+    units: 6,
+    level: 2000,
+    college: "Science",
+    semesters: ["S1"],
+    prerequisites: ["MATH1116"],
+    prerequisiteAlternatives: [["MATH1116"], ["MATH1014"]],
+    description: "Honours-track real analysis, metric spaces",
+    type: "elective",
+  },
+  MATH2322: {
+    code: "MATH2322",
+    name: "Algebra 1",
+    units: 6,
+    level: 2000,
+    college: "Science",
+    semesters: ["S2"],
+    prerequisites: ["MATH1014"],
+    prerequisiteAlternatives: [["MATH1014"], ["MATH1116"]],
+    description: "Abstract algebra, groups, rings",
+    type: "elective",
+  },
+  MATH3029: {
+    code: "MATH3029",
+    name: "Probability Modelling and Applications",
+    units: 6,
+    level: 3000,
+    college: "Science",
+    semesters: ["S1"],
+    prerequisites: ["MATH1014"],
+    description: "Stochastic processes, Markov chains, queueing theory",
+    type: "elective",
+  },
+  MATH3228: {
+    code: "MATH3228",
+    name: "Complex Analysis",
+    units: 6,
+    level: 3000,
+    college: "Science",
+    semesters: ["S1"],
+    prerequisites: ["MATH2320"],
+    prerequisiteAlternatives: [["MATH2320"], ["MATH2305"]],
+    description: "Complex functions, contour integration, residue theorem",
+    type: "elective",
+  },
+  MATH3320: {
+    code: "MATH3320",
+    name: "Advanced Analysis 2",
+    units: 6,
+    level: 3000,
+    college: "CSS",
+    semesters: ["S1"],
+    prerequisites: ["MATH2320"],
+    description:
+      "Explain the fundamental concepts of advanced analysis such as Lebesgue measure and integration and Hilbert space theory and their role in modern mathematics and applied contexts",
+    type: "elective",
+    recommended: "Improved Analysis techniques",
+  },
+  MATH3325: {
+    code: "MATH3325",
+    name: "Functional Analysis",
+    units: 6,
+    level: 3000,
+    college: "Science",
+    semesters: ["S2"],
+    prerequisites: ["MATH2320"],
+    description: "Banach and Hilbert spaces, bounded operators",
+    type: "elective",
+  },
+  MATH3511: {
+    code: "MATH3511",
+    name: "Scientific Computing",
+    units: 6,
+    level: 3000,
+    college: "Science",
+    semesters: ["S1"],
+    prerequisites: ["MATH2305"],
+    prerequisiteAlternatives: [["MATH2305"], ["MATH2405"]],
+    description:
+      "Numerical methods for ODEs and PDEs, iterative methods, stability analysis",
+    type: "elective",
+    recommended: "Core numerical methods course",
+  },
+  MATH3512: {
+    code: "MATH3512",
+    name: "Matrix Computations",
+    units: 6,
+    level: 3000,
+    college: "Science",
+    semesters: ["S2"],
+    semesterPattern: "odd_years_only",
+    offeredYears: [2025, 2027, 2029],
+    prerequisites: ["MATH2305"],
+    prerequisiteAlternatives: [["MATH2305"], ["MATH1014"]],
+    description:
+      "Numerical linear algebra, matrix factorizations, eigenvalue algorithms, SVD",
+    type: "elective",
+    recommended:
+      "Essential for large-scale simulations. OFFERED ODD YEARS ONLY.",
+  },
+  MATH3514: {
+    code: "MATH3514",
+    name: "Numerical Optimisation",
+    units: 6,
+    level: 3000,
+    college: "Science",
+    semesters: ["S2"],
+    semesterPattern: "even_years_only",
+    offeredYears: [2026, 2028, 2030],
+    prerequisites: ["MATH2305"],
+    prerequisiteAlternatives: [["MATH2305"], ["MATH1014"]],
+    description:
+      "Gradient descent, Newton methods, constrained optimization, convex optimization",
+    type: "elective",
+    recommended:
+      "Critical for antenna design and control systems. OFFERED EVEN YEARS ONLY.",
+  },
+};
