+++ src/data/courses/PHYS.ts
@@ -1,0 +1,113 @@
+import { Course } from "../../types";
+
+export const PHYS_COURSES: Record<string, Course> = {
+  PHYS1101: {
+    code: "PHYS1101",
+    name: "Physics 1",
+    units: 6,
+    level: 1000,
+    college: "Science",
+    semesters: ["S1", "S2"],
+    prerequisites: [],
+    assumedKnowledge: "HSC Physics or equivalent",
+    incompatible: ["PHYS1001"],
+    description:
+      "Mechanics, circuits, electromagnetism introduction. Recommended for engineering students.",
+    type: "foundation",
+  },
+  PHYS1201: {
+    code: "PHYS1201",
+    name: "Physics 2",
+    units: 6,
+    level: 1000,
+    college: "Science",
+    semesters: ["S2"],
+    prerequisites: ["PHYS1101"],
+    prerequisiteAlternatives: [["PHYS1101"], ["PHYS1001"]],
+    description: "Waves, optics, thermodynamics, Fourier analysis introduction",
+    type: "elective",
+  },
+  PHYS2013: {
+    code: "PHYS2013",
+    name: "Quantum Mechanics",
+    units: 6,
+    level: 2000,
+    college: "Science",
+    semesters: ["S1"],
+    prerequisites: ["PHYS1201", "MATH1014"],
+    description: "Introduction to quantum mechanics, wavefunctions, operators",
+    type: "elective",
+  },
+  PHYS2016: {
+    code: "PHYS2016",
+    name: "Electromagnetism",
+    units: 6,
+    level: 2000,
+    college: "Science",
+    semesters: ["S1"],
+    prerequisites: ["PHYS1101", "MATH1014"],
+    description:
+      "Maxwell's equations, electromagnetic waves, vector calculus applications",
+    type: "elective",
+    recommended: "Critical first-principles course for RF electronics",
+  },
+  PHYS2020: {
+    code: "PHYS2020",
+    name: "Thermal and Statistical Physics",
+    units: 6,
+    level: 2000,
+    college: "Science",
+    semesters: ["S2"],
+    prerequisites: ["PHYS1201", "MATH1014"],
+    description: "Thermodynamics, statistical mechanics",
+    type: "elective",
+  },
+  PHYS2201: {
+    code: "PHYS2201",
+    name: "Classical Mechanics",
+    units: 6,
+    level: 2000,
+    college: "Science",
+    semesters: ["S2"],
+    prerequisites: ["PHYS1101", "MATH1014"],
+    description: "Lagrangian and Hamiltonian mechanics",
+    type: "elective",
+  },
+  PHYS3032: {
+    code: "PHYS3032",
+    name: "Solid State Physics",
+    units: 6,
+    level: 3000,
+    college: "Science",
+    semesters: ["S1"],
+    prerequisites: ["PHYS2013"],
+    description:
+      "Crystal structure, band theory, semiconductors, device physics",
+    type: "elective",
+    recommended: "Essential for understanding RF component behavior",
+  },
+  PHYS3101: {
+    code: "PHYS3101",
+    name: "Advanced Quantum Mechanics",
+    units: 6,
+    level: 3000,
+    college: "Science",
+    semesters: ["S1"],
+    prerequisites: ["PHYS2013", "MATH2306"],
+    description: "Perturbation theory, scattering, identical particles",
+    type: "elective",
+  },
+  PHYS3102: {
+    code: "PHYS3102",
+    name: "Advanced Electromagnetism",
+    units: 6,
+    level: 3000,
+    college: "Science",
+    semesters: ["S2"],
+    prerequisites: ["PHYS2016", "MATH2306"],
+    description:
+      "Radiation from currents, waveguides, transmission lines, covariant formulation",
+    type: "elective",
+    recommended: "Advanced EM theory foundation for RF technologies",
+  },
+};
